PublicIP Release 1.0.0.0
The first release of PublicIP.

Description:
"PublicIP is a small 30 liner to make the public / external IP address visible to the user. This can be achieved with other Command-line tools, but I just wanted to be able to use a memorizable command that anyone could use as if it was always there. "